+++ graphql/utils.py
@@ -1,0 +1,3 @@
+def convert_snake_to_camel_case(snake: str):
+    parts = snake.split("_")
+    return "".join([x.capitalize() for x in parts])
