+++ hexa/plugins/connector_s3/models.py
@@ -1,4 +1,3 @@
-import os
 import json
 
 from django.contrib.contenttypes.models import ContentType
